backend/requirements.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
alembic>=1.7.1
python-jose[cryptography]>=3.3.0
supabase